COMP: warnings.
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=8217
git-svn-id: http://svn.slicer.org/Slicer4/trunk@8217 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/Qdec/QdecGlmDesign.cpp
@@ -785,7 +785,7 @@ int QdecGlmDesign::GenerateContrasts ( )
       if (contrast.size() != nreg )
         {
         fprintf(stderr,
-                "ERROR: QdecGlmDesign::GenerateContrasts: contrast size %d != %d\n", contrast.size(), nreg);
+                "ERROR: QdecGlmDesign::GenerateContrasts: contrast size %lu != %d\n", (unsigned long) contrast.size(), nreg);
         return -1;
         }
       contrast[nthc] = 1;
